Добавлен файл boot.php для подключения файлов. Функции вынесены в functions.php
