+++ boot.php
@@ -1,0 +1,4 @@
+<?php
+
+require_once('functions.php');
+require_once('data.php');
