+++ functions.php
@@ -16,3 +16,13 @@ function include_template($name, $data) {
 
     return $result;
 }
+
+function formatting_price($value)
+{
+    $value = ceil($value);
+    if ($value >= 1000){
+        $value = number_format($value, 0, '',' ');
+    }
+    $value .= ' &#8381;';
+    return $value;
+}
